Lab_Data_Hub-main/backend/requirements.txt
django==4.2.5
djangorestframework==3.14.0
django-filter==23.2
drf-yasg==1.21.7
djoser==2.2.0
pytest==7.4.2
pytest-django==4.5.2
psycopg2-binary==2.9.8
gunicorn==21.2.0
isort==5.12.0